+++ apps/web/src/components/properties/hooks/usePropertyPanel.ts
@@ -1,0 +1,20 @@
+import { useConsolidatedDiagramStore } from '@/stores';
+import { usePropertyForm as usePropertyFormBase } from '@repo/diagram-ui';
+
+export function usePropertyPanel<T extends Record<string, any>>(
+  entityId: string,
+  entityType: 'node' | 'arrow' | 'person',
+  initialData: T
+) {
+  const store = useConsolidatedDiagramStore();
+  
+  const updateFn = entityType === 'node' 
+    ? store.updateNodeData
+    : entityType === 'arrow' 
+    ? store.updateArrowData 
+    : store.updatePerson;
+    
+  return usePropertyFormBase(initialData, (updates) => {
+    updateFn(entityId, updates);
+  });
+}
