+++ apps/web/src/components/properties/panels/index.ts
@@ -1,0 +1,7 @@
+export { PersonJobPanelContent } from './PersonJobPanel';
+export { ConditionPanelContent } from './ConditionPanel';
+export { DBPanelContent } from './DBPanel';
+export { ArrowPanelContent } from './ArrowPanel';
+export { PersonPanelContent } from './PersonPanel';
+export { EndpointPanelContent } from './EndpointPanel';
+export { JobPanelContent } from './JobPanel';
